@huydhn 2024-02-27
@htyu I think this change is failing trunk ROCm job https://hud.pytorch.org/pytorch/pytorch/commit/78f53a3f731ee67dcffd308519ed48a745640dde.  Could you help take a look?  Due to its capacity constraint, the PR needs to have `ciflow/rocm` to run all ROCm tests, so this was missed on your PR.