@huydhn 2024-02-27
On Linux, I'm using internal devgpu.  But then ROCm running is not accessible to us too via ssh, so I guess you will still need to rely on CI to confirm